refactoring and formatting, one text change misleading past dates
