+++ app/views/casa_cases/_form.html.erb
@@ -65,7 +65,7 @@
 
       <div class="field form-group">
         <% if policy(casa_case).update_court_date? %>
-          <%= form.label :court_date, t(".next_court_date") %>
+          <%= form.label :court_date, "Add Court Date (past or future)" %>
           <br>
           <span class="datetime-year-month">
             <%= form.date_select :court_date,
@@ -166,7 +166,7 @@
             <div class="form-check">
               <%= check_box_tag "casa_case[casa_case_contact_types_attributes][][contact_type_id]",
                                 contact_type.id,
-                                @casa_case.casa_case_contact_types.map(&:contact_type_id).include?(contact_type.id),
+                                @casa_case.has_contact_type?(contact_type),
                                 id: "casa_case_contact_types_#{index}",
                                 class: ["form-check-input", "casa-case-contact-type"] %>
               <label class="form-check-label" for="casa_case_contact_types_<%= index %>">
